Move ManageSpaceActivity.kt Actions to their own file
